# chore(uv): more env var mappings