## Summary

Small follow up to https://github.com/astral-sh/uv/pull/8151

## Test Plan

Existing tests