@konstin 2024-10-15
Perf impact:

main:
```
$ hyperfine --prepare "cargo clean -p uv-cli" "cargo build -p uv-cli"
Benchmark 1: cargo build -p uv-cli
  Time (mean ± σ):      2.096 s ±  0.019 s    [User: 2.851 s, System: 0.343 s]
  Range (min … max):    2.070 s …  2.131 s    10 runs
```

PR:
```
Benchmark 1: cargo build -p uv-cli
  Time (mean ± σ):      2.093 s ±  0.044 s    [User: 2.853 s, System: 0.336 s]
  Range (min … max):    2.056 s …  2.208 s    10 runs
```

```
$ hyperfine --prepare "cargo clean -p uv-cli && cargo clean -p uv-static" "cargo build -p uv-cli"
Benchmark 1: cargo build -p uv-cli
  Time (mean ± σ):      5.169 s ±  0.044 s    [User: 6.628 s, System: 1.756 s]
  Range (min … max):    5.121 s …  5.278 s    10 runs
```

We indeed increase the dep tree, but given that uv-cli already is pretty much at the top of the build DAG, the perf impact is negligible, i.e. we build uv-static while building other workspace crates that uv-cli depends on.